requirements.txt
requests~=2.32.5
aiogram~=3.22.0
redis~=6.4.0
httpx~=0.28.1
python-dotenv~=1.1.1